Splitting the host's vim configs and plugins with the container ones.
Some of the plugins that runs on `cnvim` are not supported in my host's
`nvim` because it has an older version.

Not making the split will make me modify the `init.vim` file on each
switch between `nvim` and `cnvim` and to remove/install the different
plugins on each switch.

Therefore, I have made a copy of the `init.vim` file and the vim and
nvim plugins to the `cnvim` repo and I am mounting those to make sure I
have a clear separation.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>